% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_api.R
\name{extract_from_content}
\alias{extract_from_content}
\title{Extract content from the fetched data}
\usage{
extract_from_content(content, data_key_name, pagination_key_name)
}
\arguments{
\item{content}{a httr::content() object (return of `get_from_key`)}

\item{data_key_name}{a string the key name of data (such as "daily_quotes").}

\item{pagination_key_name}{a string or NULL the key name of pagination key. If not contained, NULL.}
}
\value{
a list (the data and pagination_key)
}
\description{
Extract content from the fetched data
}
\examples{
\dontrun{
content <- get_from_api("/prices/daily_quotes", list(date = "20201001", code = "86970"))
extract_from_content(content, "daily_quotes", "pagination_key")
}
}
