% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_method.R
\name{get_from_api}
\alias{get_from_api}
\title{Get data from J-Quants API and return the data}
\usage{
get_from_api(resource_path, query, id_token)
}
\arguments{
\item{resource_path}{a string resource_path.}

\item{query}{a list or NULL A list must be contain query parameters.}

\item{id_token}{a string id_token.}
}
\value{
a tibble or NULL.
}
\description{
Get data from J-Quants API and return the data
}
\examples{
\dontrun{
get_from_api("/listed/info", NULL, id_token)
get_from_api("/prices/daily_quotes", list(date = "20201001", code = "86970"))
}
}
