% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetcher.R
\name{get_full_data}
\alias{get_full_data}
\title{Get all data from API, considering pagination_key}
\usage{
get_full_data(resource_path, query, key_name, id_token)
}
\arguments{
\item{resource_path}{a string resource_path.}

\item{query}{a list or NULL A list must be contain query parameters.}

\item{key_name}{a string the key name of data (such as "daily_quotes").}

\item{id_token}{a string id_token.}
}
\value{
a tibble
}
\description{
Get all data from API, considering pagination_key
}
\examples{
\dontrun{
get_full_data("/listed/info", list(date="20220701"), "info", id_token)
get_full_data("/prices/daily_quotes", list(date = "20201001", code = "86970"), "daily_quotes", id_token)
}
}
