% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_market_information.R
\docType{data}
\name{market_information}
\alias{market_information}
\title{Market information code}
\format{
A tibble with 12 rows 2 columns:
\itemize{
  \item{MarketType}: {Market type}
  \item{MarketCode}: {The same as a "MarketCode" column in a return tibble of "get_info()"}
}
}
\source{
\url{https://jpx.gitbook.io/j-quants-api/api-reference/listed-api/segment}
}
\usage{
market_information
}
\description{
A tibble of market type and market code
which is the same as a "MarketCode" column in a return tibble of "get_info()".
}
\examples{
\dontrun{
market_information
}
}
\keyword{datasets}
