% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_from_api.R
\name{get_full_data}
\alias{get_full_data}
\title{Get all data from API, considering pagination_key}
\usage{
get_full_data(
  resource_path,
  query,
  data_key_name,
  pagination_key_name,
  is_paginationable,
  id_token
)
}
\arguments{
\item{resource_path}{a string resource_path.}

\item{query}{a list or NULL A list must be contain query parameters.}

\item{data_key_name}{a string the key name of data (such as "daily_quotes").}

\item{pagination_key_name}{a string or NULL the key name of pagination key. If not contained, NULL.}

\item{is_paginationable}{a bool whether the data from the resource path can have pagination key.}

\item{id_token}{a string id_token.}
}
\value{
a tibble
}
\description{
Get all data from API, considering pagination_key
}
\examples{
\dontrun{
get_full_data("/listed/info", list(date="20220701"), "info", "pagination_key", FALSE, id_token)
get_full_data("/prices/daily_quotes", list(date = "20201001", code = "86970"), "daily_quotes", "pagination_key", TRUE, id_token)
}
}
