% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_info}
\alias{get_info}
\alias{get_sections}
\alias{get_daily_quotes}
\alias{get_financial_statements}
\alias{get_financial_announcement}
\title{Get data from J-Quants API}
\usage{
get_info(code, id_token = Sys.getenv("JQUANTSR_ID_TOKEN"))

get_sections(id_token = Sys.getenv("JQUANTSR_ID_TOKEN"))

get_daily_quotes(
  code,
  from,
  to,
  date,
  id_token = Sys.getenv("JQUANTSR_ID_TOKEN")
)

get_financial_statements(
  code,
  date,
  id_token = Sys.getenv("JQUANTSR_ID_TOKEN")
)

get_financial_announcement(id_token = Sys.getenv("JQUANTSR_ID_TOKEN"))
}
\arguments{
\item{code}{a string stock code.}

\item{id_token}{a string your id_token.}

\item{from}{a string the start date of data. "\%Y\%m\%d" and "\%Y-\%m-\%d" formats are valid.}

\item{to}{a string the end date of data. "\%Y\%m\%d" and "\%Y-\%m-\%d" formats are valid.}

\item{date}{a string the date of data. "\%Y\%m\%d" and "\%Y-\%m-\%d" formats are valid.}
}
\value{
if successfully fetched data (= status code of the API is 200), a tibble of the fetched data.
if not successfully (= status code of the API is not 200), none and message from the API is printed.
}
\description{
Get data from J-Quants API.

When you have run "authorize()", you don't need to set "id_token" parameter,
because "authorize()" set your ID token in a environment variable named "JQUANTSR_ID_TOKEN".
}
\details{
parameter "code", "from", "to", "date" are required as follows.
\itemize{
  \item{"get_info()"}: {"code" is optional.}
  \item{"get_daily_quotes()"}: {At least one of "code" and "date" are mandatory. If "from" and "to" parameters are used, "code" is mandatory.}
  \item{"get_financial_annoucements()"}: {At least one of "code" and "date" are mandatory.}
}
For more information, see \href{https://jpx.gitbook.io/j-quants-api/api-reference}{J-Quants API reference}.
}
\examples{
\dontrun{
# you must have done "authorize()" before running "get_info" or other functions,
authorize(refresh_token = YOUR_REFRESH_TOKEN)
get_info()
get_info(code = "86970")
get_sections()
get_daily_quotes(date = "20220701")
get_daily_quotes(code = "86970", from = "20220101", to = "20220630")
get_financial_statements(code = "86970")
get_financial_statements(date = "20220105")
get_financial_annoucement()
}
}
